Create README.md
This commit creates the initial version of the README.md file for my GitHub profile. It includes an overview of my professional experience in SAP Business One, SQL, and data analysis, as well as my career aspirations to become a Business Intelligence Analyst. It highlights my skills in Python, SQL, data visualization, and web scraping, along with key projects that demonstrate my expertise in financial analysis, data reporting, and application development. The README serves as a comprehensive introduction to my background and goals for potential collaborators and employers